+++ Userland/userCodeModule/include/userlib.h
@@ -216,7 +216,7 @@ void beep(uint32_t frequency, int duration);
 // Memory
 //================================================================================================================================
 
-void * my_malloc(uint16_t size);
+void * my_malloc(uint32_t size);
 
 void my_free(void * addr_to_free);
